Make SyncClock instance; add time source prefixes
Convert SyncClock from a static utility to an instance-based class to avoid global/static state and enable per-TrustedTime instances (constructor, update, elapsedSinceAnchorMs, lastUptimeMs/lastWallMs, dispose). Wire a SyncClock instance into TrustedTimeImpl and update usages/tests accordingly. Add TimeSource prefix constants (ntp:, nts:, https:) and use them in NTP/NTS/HTTPS sources. Improve NtsSource._generateUniqueId to use cryptographically secure randomness. Fix a race in IntegrityMonitor._emit by catching add() errors when the stream closes. Make confidenceScore math more stable by using toDouble() and add documentation about numeric safety. Update tests to reflect instance-based SyncClock and dispose behavior.

+++ lib/src/models.dart
@@ -100,12 +100,16 @@ final class TrustAnchor {
   ///
   /// As the score approaches 0.1 (the absolute floor), applications should 
   /// consider the anchor "stale" and trigger a resync.
+  ///
+  /// **Safe Range**: This calculation uses double-precision arithmetic and 
+  /// [Duration.inMinutes] (returning a 64-bit int). It is numerically stable 
+  /// for anchor ages exceeding 4,000 years.
   double get confidenceScore {
     final age = DateTime.now().difference(syncTime);
     
     // Freshness Decay (Half-life model)
     final halfLifeHours = confidence == ConfidenceLevel.high ? 6.0 : 2.0;
-    final decay = age.inMinutes / (halfLifeHours * 60);
+    final decay = age.inMinutes.toDouble() / (halfLifeHours * 60.0);
     final freshness = 1.0 / (1.0 + decay);
 
     // Base Trust Floor
